addressing new pr comments
