+++ Cargo.toml
@@ -179,6 +179,7 @@ parquet = { version = "58.1.0", default-features = false, features = [
     "object_store",
 ] }
 pbjson = { version = "0.9.0" }
+pin-project = "1"
 pbjson-types = "0.9"
 # Should match arrow-flight's version of prost.
 prost = "0.14.1"
