+++ datafusion/datasource-parquet/Cargo.toml
@@ -53,7 +53,6 @@ log = { workspace = true }
 object_store = { workspace = true }
 parking_lot = { workspace = true }
 parquet = { workspace = true }
-pin-project = "1"
 tokio = { workspace = true }
 
 [dev-dependencies]
