+++ datafusion/physical-expr-common/Cargo.toml
@@ -49,6 +49,7 @@ hashbrown = { workspace = true }
 indexmap = { workspace = true }
 itertools = { workspace = true }
 parking_lot = { workspace = true }
+pin-project = { workspace = true }
 
 [dev-dependencies]
 criterion = { workspace = true }
